requirements.txt
# requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
webdriver-manager==4.8.0
